Run HOL4/bin/build in Makefile; document.
The HOL4 INSTALL file suggests that the user should run the HOL4/bin/build
script to finish the HOL4 install process. This was the slowest of the steps
done during the setup phase. This change suggests that the user not run build
manually and calls it from the Makefile, thus moving the CPU time into the
batch phase.