loop_bounds: more principle for preemptionPoint
We've been manually adding bounds for functions which call preemptionPoint, but
that means manually checking that we've found them all, and for larger numbers
of runs I'm starting to make mistakes.  Add a mechanism to quickly check this.
Has to be configured to the name of the functions limited. Also allows pruning
functions considered unreachable in this run.