Fix previous; include a compatibility workaround.
Fixes a bug in previous.

Also adds a hook whereby a target can request a previous level of
compatibility with functions that contain ASM statements and other
non-calling-convention compatible functions that may dangerously
adjust the stack pointer.

(The previous level of compatibility is actually unsound, because
it assumed that e.g. the fastpath restore assembly block did not
do what it did.)