The calls-on-memory optimisation.
Track, for each memory object, the number of calls to various
functions that have gone into it. Well, a range, min-max. Avoid
producing function relations for function call pairs that can't
reasonably be related. This may substantially reduce the problem
complexity for function calls in loops that get unrolled a fair
bit.