Fix bug in shiftl offset logic.
This one was a howler, but there might be more subtle offset problems
around.

Frustratingly, the obvious way to test the offset logic won't work.
Any stack access, particularly around a function call, requires a
serious correctness argument to line up the way the analysis sees it
with what might happen in the general case.