Take more care with loop offsets in multi-entry.
Normally the head point is splittable and the first node of the loop
to be visited, but in the multi-entry case, the head point may not be
the first point visited in one of the cases. This may create an off-by-1
issue in the generated proof script, and needs to be handled a little
more carefully.