Change side for rodata proofs.
Of course, it's much easier to prove the C output memory is unchanged,
since the pvalid arguments directly relate the C memory access addresses
to the reserved .rodata addresses. I think the reason it was done
differently before is that there must always be an input asm-side memory,
whereas the C one may be missing, and we still need to know rodata.

New version: do everything on C side if possible. If only ASM side exists,
assume rodata holds. So sometimes read-only functions will happen after
ASM updates and the solvers will have to work hard to figure out that the
updates (on the ASM side) were protected (on the C side). But most of the
time rodata will prove quickly because it's easy to see the updates aren't
touching the witness address.