Probable solution to the map_kernel_window saga.
The key monster here is a function with three loops, each of a known
iteration count, where the count in the second loop is 16 and it is
unrolled. That doesn't get caught by any of our existing detectors.

The linear sequence mechanism can be used to detect what's happened. The
interesting elements of loop 2 in the C side are matched by elements outside
all loops on the binary side. That can be used to trigger a loop bound check.