Support some more instructions and stack cases.
Firstly, handle some more instruction forms that are showing up
in the source & binary.

Secondly, adjust the way that conditional path comparisons are
handled in stack_logic. This whole calculation is mostly an
over-precise way of trying to work around the presence of
non-returning function calls. Downstream from these we compute
stack offsets which are conditional on the impossible path, which
then need to be simplified. This new formulation seems to handle
the new cases, but may need to be strengthened yet again.