New directory of scripts.
In particular, move some setup code for making sure HOL4 is running
from seL4-example/configure_default.sh into its own script.

Also copy in a debug script, which I've had privately for a long
while now. It's a collection of snippets which are useful to quote
to the interactive python interpreter to diagnose various failures.