word64: fix cancelAllIPCs regression
We fix a regression affecting the function `cancelAllIPCs`, which
went from `True` to `ProofEXCEPT` after `commit f64510...` made
`mk_cast` part of the architecture specification.

The problem was an out-of-order check: cast should be a no-op on
non-Word kinds, but the rv64 cast implementation tried to check
the length of words before checking that the kinds are actually
Word kinds.

The patch gets rid of `mk_cast_generic`, since the relevant checks
are now done on a per-architecture basis.

Signed-off-by: Zoltan A. Kocsis <zoltan.kocsis@data61.csiro.au>