word64: future-proof zero-wired register check
Zero-wired registers always contain the value 0. As such, writes
to zero-wired registers have no effect. Handling this requires a
bit of special logic, which would not work if the name of a zero-
wired register could occur as the prefix of the name of another,
non-zero-wired register (e.g. on an architecture where r1 is a
zero-wired register, while r10 is not).

This patch fixes the prefix issue. While none of our current
architectures exhibit the phenomenon, this should help future-
proof our code.

Signed-off-by: Zoltan A. Kocsis <zoltan.kocsis@data61.csiro.au>