word64: make cast_pair arch-specific, restore orig. mk_cast behavior
The rv64 calling convention maintains the invariant that all 32-bit
values, including unsigned ones, are held in sign-extended format in
64-bit registers.

From `2069ea7...` to now, the implementation of `syntax.arch.mk_cast`
for rv64 always performed sign extensions for 32-to-64-bit casts,
even when rendering unsigned casts in SydTV-GL.

The intended behavior for the October calling convention update was
that the sign extension should only be applied for the two
`cast_pair` calls in `mk_eqs`, since that's where this aspect of the
calling convention is relevant. This patch reverts the behavior
that was broken by `2069ea7...`.

Signed-off-by: Zoltan A. Kocsis <zoltan.kocsis@data61.csiro.au>