Refactor write_solv_script
The function `write_solv_script` now writes a given solver script
to a temporary file, then moves it into its permanent location,
basing the permanent filename on the sha256sum of the written file.

Previously, the permanent filename was calculated based on the hash
of the input messages, not the hash of the written file.

The change fixes a problem where two files with different preambles
but identical solver scripts would compete for the same permanent
location.

Signed-off-by: Zoltan Kocsis <zoltan.kocsis@data61.csiro.au>