Update seL4 target Makefile (cleanup, automate manual fix-ups)
Previously, when the Makefile was used to build a RISC-V seL4 target, the
resulting target files were not immediately usable by `graph-refine` itself.
The following tasks had to be performed manually:
 - AUIPC fix-up: using the check-addr.py script in tools/
 - `sfence_vma` to `sfence.vma` instruction renaming: manually
 - `StackBounds.txt` calculations: using the `stack_bounds.py` script

These tasks are now automated as part of the Makefile.

The Makefile still contains some architecture-specific (RISC-V only) code,
which is now commented with FIXMEs.

Signed-off-by: Zoltan Kocsis <zoltan.kocsis@data61.csiro.au>