Improve model behavior in `start_prove_task_with_solver`
Previously, `start_prove_task_with_solver` passed its `model` argument around
directly. The correctness of this behavior depends upon the implementation
details of `solver.Solver.check_model_iteration` to ensure that such a model is
not updated directly.

We change the behavior, so that we pass separate copies of the `model` instead,
which results in correct behavior independly of `check_model_iteration`.

Signed-off-by: Zoltan Kocsis <zoltan.kocsis@data61.csiro.au>