add a simple parallel job runner
Add a simple graph-refine job queue that runs jobs serially, but runs
functions within a job in parallel.

In the Nix/Docker build, add support for dispatching to different
graph-refine versions for different architectures. This is a temporary
measure while the ARM and RISCV64 branches are not yet integrated.

Also modify the `build` workflow to build graph-refine Docker images
including both ARM and RISCV64 versions.

Signed-off-by: Matthew Brecknell <matt@kry10.com>