fix calling convention preconditions
Before this fix, the preconditions were too weak. In particular, in the
case of a function which returns a large value on the stack, they
allowed for a return value region in high memory that wraps around 0,
potentially clobbering a stack in low memory.

This fix strengthens the preconditions to forbid this.

Signed-off-by: Nick Spinale <nick@nickspinale.com>